Module Tp1Concurrentes::body::process_data

Functions

columns_index 🔒

A function that takes a vector of headers and returns a Columns struct with the indices of the relevant columns. This is used to determine the indices of the columns in the CSV files in case the columns are out of order.

gather_data

Function that processes the CSV files and gathers the data.

merge_game_maps 🔒

Function that merges two HashMaps of TopGames.

merge_lang_maps 🔒

Function that merges two HashMaps of TopLanguages.